create displayTime helper
